consultationDetail component broken down
